refactor(helm): split operation.go into execute.go + helmfile.go
Match the one-function-per-file convention used in the rest of the
helm/ module (and sibling modules like kcl/, argocd/).

- operation.go -> helmfile.go (HelmfileOperation)
- New execute.go (Execute)
- New util.go (splitValues helper, shared between the two)

No behavioral change. Validated with task test-helm.

Refs #239
